# backend/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Reflow Gateway API
  description: |
    MCP (Model Context Protocol) multiplexing gateway. Provides authentication,
    fine-grained authorization, credential injection, and transport abstraction
    for upstream MCP servers.

    ## Authentication

    All protected endpoints require a JWT token via `Authorization: Bearer <token>`.
    Obtain a token by registering and logging in, or by creating an API token.

    ## Authorization Model

    The gateway operates on a **default-deny** model. Access to MCP servers,
    tools, resources, and prompts requires explicit authorization policies.

    ## MCP Protocol

    The gateway exposes a single MCP endpoint at `/mcp` supporting Streamable HTTP
    (POST for JSON-RPC, GET for SSE notifications, DELETE to close sessions).
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/policies:
    get:
      tags: [Policies]
      summary: List all policies
      operationId: listPolicies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of authorization policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorizationPolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Policies]
      summary: Create policy
      description: |
        Creates a new authorization policy. Admin only.
        Policies are evaluated by priority (highest first). First match wins.
        If no policy matches, access is denied (default-deny).
      operationId: createPolicy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationPolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/policies/{id}:
    get:
      tags: [Policies]
      summary: Get policy
      operationId: getPolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Policy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationPolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Policies]
      summary: Update policy
      description: Admin only. All fields are optional (partial update).
      operationId: updatePolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyRequest"
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationPolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Policies]
      summary: Delete policy
      description: Admin only.
      operationId: deletePolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Policy deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/policies/{id}/subjects:
    post:
      tags: [Policies]
      summary: Add subject to policy
      description: Admin only.
      operationId: addPolicySubject
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectRequest"
      responses:
        "201":
          description: Subject added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicySubject"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/policies/{id}/subjects/{subjectId}:
    delete:
      tags: [Policies]
      summary: Remove subject from policy
      description: Admin only.
      operationId: deletePolicySubject
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Subject removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ──────────────────────── Environment Config ────────────────────────
  /api/targets/{id}/env:
    get:
      tags: [Environment Config]
      summary: List all env configs
      description: Lists all environment configurations for a target, grouped by scope.
      operationId: listEnvConfigs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Environment configuration list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetEnvConfigList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/targets/{id}/env/resolve:
    get:
      tags: [Environment Config]
      summary: Resolve env configs for current user
      description: |
        Returns the fully resolved environment configuration for the current user,
        applying the resolution priority: user > group > role > default.
      operationId: resolveEnvConfigs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Resolved configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedEnvConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AuthorizationPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        target_id:
          type: string
          format: uuid
          nullable: true
          description: "null = applies to all targets"
        resource_type:
          type: string
          enum: [all, tool, resource, prompt]
        resource_pattern:
          type: string
          nullable: true
          description: Regex pattern for resource name matching
        effect:
          type: string
          enum: [allow, deny]
        priority:
          type: integer
          description: Higher values are evaluated first
        enabled:
          type: boolean
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/PolicySubject"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PolicySubject:
      type: object
      properties:
        id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        subject_type:
          type: string
          enum: [user, role, group, everyone]
        subject_value:
          type: string
          nullable: true
          description: "user_id, role name, or group name. null for 'everyone'"
        created_at:
          type: string
          format: date-time
...
    TargetEnvConfigList:
      type: object
      properties:
        target_id:
          type: string
          format: uuid
        target_name:
          type: string
        default_configs:
          type: array
          items:
            $ref: "#/components/schemas/EnvConfigListItem"
        role_configs:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/EnvConfigListItem"
        group_configs:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/EnvConfigListItem"
        user_configs:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/EnvConfigListItem"
    EnvConfigListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        description:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
    ResolvedEnvConfig:
      type: object
      properties:
        target_id:
          type: string
          format: uuid
        target_name:
          type: string
        configs:
          type: array
          items:
            $ref: "#/components/schemas/EnvConfigInfo"
    EnvConfigInfo:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        source:
          type: string
          enum: [user, group, role, default]
        scope_value:
          type: string
        description:
          type: string
...
    CreatePolicyRequest:
      type: object
      required: [name, resource_type, effect, priority]
      properties:
        name:
          type: string
        description:
          type: string
        target_id:
          type: string
          format: uuid
          nullable: true
        resource_type:
          type: string
          enum: [all, tool, resource, prompt]
        resource_pattern:
          type: string
          nullable: true
        effect:
          type: string
          enum: [allow, deny]
        priority:
          type: integer
        enabled:
          type: boolean
          default: true
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/CreateSubjectRequest"
    UpdatePolicyRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        target_id:
          type: string
          format: uuid
        resource_type:
          type: string
          enum: [all, tool, resource, prompt]
        resource_pattern:
          type: string
        effect:
          type: string
          enum: [allow, deny]
        priority:
          type: integer
        enabled:
          type: boolean
    CreateSubjectRequest:
      type: object
      required: [subject_type]
      properties:
        subject_type:
          type: string
          enum: [user, role, group, everyone]
        subject_value:
          type: string
          nullable: true
          description: "user_id, role name, or group name. Omit for 'everyone'"
...